crowd

+++ src/components/statesModal/crowdStatesModal.ts
@@ -1,12 +1,11 @@
 import Vue from "vue";
 import Component from "vue-class-component";
 import { IProjectableModel, Projector } from "vue-mf-module/dist/helpers/Projector";
-import { Prop, Watch } from "vue-property-decorator";
-import ValidateDirective from 'vue-mf-module';
+import { Prop } from "vue-property-decorator";
+import ValidateDirective, { MessageService } from 'vue-mf-module';
+import { statesService } from "@/services/statesService";
 import { Drag, Drop } from 'vue-drag-drop';
 import { store } from "@/store";
-import { statesService } from "@/services/statesService";
-import { CONFIGURATION } from "@/configuration";
 
 @Component({
   directives: {
@@ -15,39 +14,36 @@ import { CONFIGURATION } from "@/configuration";
     Drag, Drop
   }
 })
-export default class crowdStatesModal extends Vue {
+export default class statesModal extends Vue {
+
   @Prop()
   value!: IProjectableModel<server.Group>;
 
-  get states(): server.State[] {
-    return Array.from(store.getters.crowdplanning.getStates(this.value.data.id) || []);
-  }
+  states: server.State[] = [];
 
   get sortedStates() {
-    return this.states.sort((a, b) => a.orderIndex - b.orderIndex);
+    return Array.from(this.states).sort((a, b) => a.orderIndex - b.orderIndex);
   }
 
-  async mounted() { }
+  async mounted() {
+    this.states = Array.from(await statesService.getStates(this.value.data));
+  }
 
   handleDrop(to: server.State, from: server.State) {
-    var temp = Array.from(this.sortedStates);
-    temp.splice(temp.indexOf(from), 1);
-    temp.splice(temp.indexOf(to), 0, from);
-    this.recalcIndex(temp);
+    this.states.splice(this.states.indexOf(from), 1);
+    this.states.splice(this.states.indexOf(to), 0, from);
+    this.recalcIndex();
   }
 
-  handleDropState(generalStatus: string, from: server.State) {
-    if (generalStatus != from.generalStatus) {
-      from.generalStatus = generalStatus;
-
-      store.actions.crowdplanning.setState({ state: from, groupId: this.value.data.id });
-    }
+  handleDropState(state: string, from: server.State) {
+    if (state != from.state)
+      from.state = state;
   }
 
-  recalcIndex(list: server.State[]) {
-    for (let index = 0; index < list.length; index++) {
-      const element = list[index];
-      switch (element.generalStatus) {
+  recalcIndex() {
+    for (let index = 0; index < this.states.length; index++) {
+      const element = this.states[index];
+      switch (element.state) {
         case 'New': element.orderIndex = 100 + index + 1; break;
         case 'Open': element.orderIndex = 200 + index + 1; break;
         case 'Active': element.orderIndex = 300 + index + 1; break;
@@ -57,8 +53,18 @@ export default class crowdStatesModal extends Vue {
     }
   }
 
-  async close() {
-    for (const state of this.states) {
+  close() {
+    this.value.resolve(this.value.data)
+  }
+
+  async confirm() {
+
+    if (this.states.some(s => !s.name || !s.shortName)) {
+      MessageService.Instance.send("ERROR", this.$t("states.set-error", "Riempire tutti i campi"));
+      return;
+    }
+
+    for (const state of Array.from(this.states)) {
       await statesService.setState(state, this.value.data.id);
     }
 
@@ -66,23 +72,21 @@ export default class crowdStatesModal extends Vue {
   }
 
   async remove(s: server.State) {
+    var idx = this.states.indexOf(s);
+    if (idx >= 0)
+      this.states.splice(idx, 1);
 
-    var states = Array.from(this.states);
-    var idx = states.indexOf(s);
-    if (idx >= 0) states.splice(idx, 1);
-
-    store.actions.crowdplanning.setStates({ groupId: this.value.data.id, states });
-    await statesService.removeState(s.id);
+    if (s.id) {
+      await statesService.removeState(s.id);
+      store.actions.crowdplanning.setStates({ groupId: this.value.data.id, states: this.states });
+    }
   }
 
   async addState(s: string) {
-    var states = Array.from(this.states);
-    states.push({ generalStatus: s } as server.State)
-    this.recalcIndex(states);
-    store.actions.crowdplanning.setStates({ groupId: this.value.data.id, states });
-  }
-
-  hasPermission(permission: string): boolean {
-    return this.$can(`${CONFIGURATION.context}.${permission}`);
+    this.states.push({
+      generalStatus: s,    
+      reference: this.value.data.reference
+    } as unknown as server.State);
+    this.recalcIndex();
   }
 }
